重写数据库的命名

+++ src/main/java/com/soft1841/cn/dao/GoodsDAO.java
@@ -33,7 +33,7 @@ public interface GoodsDAO {
      * @param id
      * @return
      */
-    Entity getGoodsById(int id) throws SQLException;
+    Entity getGoodsById(long id) throws SQLException;
 
     /**
      * 修改商品信息
